% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{age_count_for_year}
\alias{age_count_for_year}
\title{Count ages for all years}
\usage{
age_count_for_year(data)
}
\arguments{
\item{data}{A data frame with at least columns \code{year} and \code{age}.}
}
\value{
A tibble with columns: year, age, n
}
\description{
Count ages for all years
}
