% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_summary_for_year}
\alias{age_summary_for_year}
\title{Makes a summary of max and mean ages for all years.}
\usage{
age_summary_for_year(data)
}
\arguments{
\item{data}{A data frame with at least columns \code{year} and \code{age}.}
}
\value{
A tibble with columns: year, age, n
}
\description{
Makes a summary of max and mean ages for all years.
}
