% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleaning_herring}
\alias{cleaning_herring}
\title{Cleaning raw data.}
\usage{
cleaning_herring(data)
}
\arguments{
\item{data}{raw fish (NSSH) data.}
}
\value{
A \code{data.frame}
}
\description{
Function to clean raw data from missing values in columns with biological
data.
}
