% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{filter_ocean_points}
\alias{filter_ocean_points}
\title{Filters out herring data which is within the normal NSSH distribution area.}
\usage{
filter_ocean_points(data)
}
\arguments{
\item{data}{The map ready dataset with colums lat and lon.}
}
\value{
A tibble with same columns as input, filtered for correct lat/lon.
}
\description{
Filters out herring data which is within the normal NSSH distribution area.
}
