% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_models.R
\name{gompertz_model}
\alias{gompertz_model}
\title{Gompertz growth model.}
\usage{
gompertz_model(t, L_inf, k, a)
}
\arguments{
\item{t}{Age.}

\item{L_inf}{Asymptotic length.}

\item{k}{Growth rate.}

\item{a}{Shape parameter.}
}
\value{
Predicted length of NSSH.
}
\description{
Gompertz growth model.
}
