% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{herring_read}
\alias{herring_read}
\title{Reads IMR NSSH file}
\usage{
herring_read()
}
\arguments{
\item{path}{Path to the csv file.}
}
\value{
A \code{data.frame}
}
\description{
This function is to import the dataset from IMR's database.
}
