% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{location_catches_summary}
\alias{location_catches_summary}
\title{Count number of unique fish IDs per year per location (lat/lon) and adding
filter for area of interest.}
\usage{
location_catches_summary(data)
}
\arguments{
\item{data}{Cleaned herring dataset.}
}
\value{
A tibble with columns year, lon, lat, n_fish and ids.
}
\description{
Count number of unique fish IDs per year per location (lat/lon) and adding
filter for area of interest.
}
