% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_models.R
\name{logistic_model}
\alias{logistic_model}
\title{Logistic growth model}
\usage{
logistic_model(t, Linf, k, t0)
}
\arguments{
\item{t}{Age (numeric vector)}

\item{Linf}{Asymptotic length}

\item{k}{Growth rate}

\item{t0}{Inflection parameter}
}
\value{
Length-at-age
}
\description{
Logistic growth model
}
