% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_models.R
\name{vbgm}
\alias{vbgm}
\title{Von Bertalanffy Growth Model.}
\usage{
vbgm(t, L_inf, k, t0)
}
\arguments{
\item{t}{Age.}

\item{L_inf}{Asymptotic length.}

\item{k}{Growth coefficient.}

\item{t0}{Theoretical age at length zero.}
}
\value{
Predicted length of NSSH.
}
\description{
Von Bertalanffy Growth Model.
}
