% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{weight_per_year}
\alias{weight_per_year}
\title{Count total weight (tonnes) of fish per year.}
\usage{
weight_per_year(data)
}
\arguments{
\item{data}{Cleaned herring dataset.}
}
\value{
A tibble with columns year and total_weight.
}
\description{
Count total weight (tonnes) of fish per year.
}
