% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_count_for_year}
\alias{age_count_for_year}
\title{Age count by year.}
\usage{
age_count_for_year(data)
}
\arguments{
\item{data}{A data frame with at least columns \code{year} and \code{age}.}
}
\value{
A tibble with columns: year, age, n
}
\description{
Counts number of fish per age for all years.
}
