% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{cleaning_herring}
\alias{cleaning_herring}
\title{Cleans the NSSH file.}
\usage{
cleaning_herring(data)
}
\arguments{
\item{data}{raw NSSH data.}
}
\value{
A \code{data.frame}
}
\description{
This function is used to clean raw data from missing values, and converting
to numerical values.
}
