% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_per_year}
\alias{count_per_year}
\title{Count fish per year using unique (id, indno) pairs.}
\usage{
count_per_year(data)
}
\arguments{
\item{data}{A cleaned herring dataset.}
}
\value{
A tibble with columns year and n_ids.
}
\description{
Count fish per year using unique (id, indno) pairs.
}
