% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{herring_read}
\alias{herring_read}
\title{Reads IMR NSSH file}
\usage{
herring_read()
}
\value{
A \code{data.frame}
}
\description{
Function to import the raw NSSH dataset from the
Institute of Marine Research database.
}
