% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{location_catches_summary}
\alias{location_catches_summary}
\title{Location catches summary.}
\usage{
location_catches_summary(data)
}
\arguments{
\item{data}{A data frame with at least columns \code{year}, \code{month}, \code{lat}, \code{lon}, \code{age} and \code{weight}.}
}
\value{
A tibble with columns year, month, lat, lon, n_fish, mean_age and mean_weight.
}
\description{
Counts number of fish per year per location (lat/lon), and calculates
mean age and weight. Additionally adding a filter for area of
interest of NSSH.
}
