% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_nssh_data}
\alias{process_nssh_data}
\title{Process pipeline.}
\usage{
process_nssh_data()
}
\value{
A list with processed datasets.
}
\description{
This function is used to process NSSH Data for Shiny App.
Loads raw data and applies all processing steps (whole pipeline)
provided by underlying NSSH package functions.
}
