% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weight_per_year}
\alias{weight_per_year}
\title{Total weight per year.}
\usage{
weight_per_year(data)
}
\arguments{
\item{data}{A data frame with \code{year} and individual \code{weight}.}
}
\value{
A tibble with columns year and total_weight.
}
\description{
Counts total weight (tonnes) of fish per year.
}
